+++ src/shared/utilities/stopPropagation.ts
@@ -1,0 +1,6 @@
+/**
+ * Helper method to stop propagation
+ * @param e {React.MouseEvent}
+ */
+export const stopPropagation = (e: React.MouseEvent<any>) =>
+    e.stopPropagation();
